MULE-10531: Migrate HTTP test that needs to access plugin classes from IT test to Functional Test Case in HTTP Extension - Moved tests to http extension in order to unignore them and allow access to internal extension classes from test
