+++ extensions/http/src/test/resources/http-request-functional-config.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<mule xmlns="http://www.mulesoft.org/schema/mule/core"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xmlns:httpn="http://www.mulesoft.org/schema/mule/httpn"
+      xsi:schemaLocation="
+               http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd
+               http://www.mulesoft.org/schema/mule/httpn http://www.mulesoft.org/schema/mule/httpn/current/mule-httpn.xsd">
+
+    <httpn:request-config name="requestConfig" basePath="basePath">
+        <httpn:request-connection host="localhost" port="${httpPort}"/>
+    </httpn:request-config>
+
+    <httpn:request-config name="requestConfigHttp">
+        <httpn:request-connection host="localhost"/>
+    </httpn:request-config>
+
+
+    <httpn:request-config name="requestConfigHttps">
+        <httpn:request-connection protocol="HTTPS" host="localhost"/>
+    </httpn:request-config>
+
+    <flow name="requestFlow">
+        <httpn:request config-ref="requestConfig" method="POST" path="requestPath"  key="ANY"/>
+    </flow>
+
+</mule>
