+++ tests/http/src/test/java/org/mule/test/module/http/functional/requester/HttpRequestFunctionalTestCase.java
@@ -7,23 +7,14 @@
 package org.mule.test.module.http.functional.requester;
 
 import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.hasEntry;
-import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsNull.nullValue;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.mule.runtime.module.http.api.HttpConstants.HttpStatus.OK;
-import static org.mule.runtime.module.http.api.HttpConstants.Protocols.HTTP;
-import static org.mule.runtime.module.http.api.HttpConstants.Protocols.HTTPS;
-import static org.mule.test.module.extension.internal.util.ExtensionsTestUtils.getConfigurationInstanceFromRegistry;
 import static org.mule.test.module.http.functional.matcher.HttpMessageAttributesMatchers.hasStatusCode;
-
 import org.mule.extension.http.api.HttpResponseAttributes;
-import org.mule.extension.http.internal.request.validator.HttpRequesterProvider;
 import org.mule.runtime.core.api.Event;
-import org.mule.runtime.core.internal.connection.ConnectionProviderWrapper;
-import org.mule.runtime.extension.api.runtime.ConfigurationInstance;
 import org.mule.tck.junit4.rule.DynamicPort;
 
 import java.io.IOException;
@@ -33,7 +24,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.eclipse.jetty.server.Request;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -41,42 +31,10 @@ public class HttpRequestFunctionalTestCase extends AbstractHttpRequestTestCase {
 
   private static final String TEST_HEADER_NAME = "TestHeaderName";
   private static final String TEST_HEADER_VALUE = "TestHeaderValue";
-  private static final String DEFAULT_PORT_HTTP_REQUEST_CONFIG_NAME = "requestConfigHttp";
-  private static final String DEFAULT_PORT_HTTPS_REQUEST_CONFIG_NAME = "requestConfigHttps";
 
   @Override
   protected String getConfigFile() {
     return "http-request-functional-config.xml";
-  }
-
-  @Ignore("MULE-10531")
-  @Test
-  public void requestConfigDefaultPortHttp() throws Exception {
-    ConfigurationInstance config =
-        getConfigurationInstanceFromRegistry(DEFAULT_PORT_HTTP_REQUEST_CONFIG_NAME, testEvent(), muleContext);
-    ConnectionProviderWrapper providerWrapper = (ConnectionProviderWrapper) config.getConnectionProvider().get();
-    HttpRequesterProvider provider = (HttpRequesterProvider) providerWrapper.getDelegate();
-    assertThat(provider.getPort().apply(testEvent()), is(HTTP.getDefaultPort()));
-  }
-
-  @Ignore("MULE-10531")
-  @Test
-  public void requestConfigDefaultPortHttps() throws Exception {
-    ConfigurationInstance config =
-        getConfigurationInstanceFromRegistry(DEFAULT_PORT_HTTPS_REQUEST_CONFIG_NAME, testEvent(), muleContext);
-    ConnectionProviderWrapper providerWrapper = (ConnectionProviderWrapper) config.getConnectionProvider().get();
-    HttpRequesterProvider provider = (HttpRequesterProvider) providerWrapper.getDelegate();
-    assertThat(provider.getPort().apply(testEvent()), is(HTTPS.getDefaultPort()));
-  }
-
-  @Ignore("MULE-10531")
-  @Test
-  public void requestConfigDefaultTlsContextHttps() throws Exception {
-    ConfigurationInstance config =
-        getConfigurationInstanceFromRegistry(DEFAULT_PORT_HTTPS_REQUEST_CONFIG_NAME, testEvent(), muleContext);
-    ConnectionProviderWrapper providerWrapper = (ConnectionProviderWrapper) config.getConnectionProvider().get();
-    HttpRequesterProvider provider = (HttpRequesterProvider) providerWrapper.getDelegate();
-    assertThat(provider.getTlsContext(), notNullValue());
   }
 
   @Test
